♻️ Refactor: #136 스타일 컴포넌트 일원화
