--- src/pages/NotFound/NotFoundStyle.jsx
@@ -1,17 +1,0 @@
-import styled from 'styled-components';
-
-const Image = styled.img`
-  margin-top: 30vh;
-`;
-
-const NotFoundText = styled.p`
-  font-size: 16px;
-`;
-const NotFoundBtn = styled.button`
-  border: none;
-  border-radius: 20px;
-  background-color: var(--basic-yellow);
-  padding: 10px 20px;
-`;
-
-export { Image, NotFoundText, NotFoundBtn };
