+++ src/pages/SearchPage/SearchPage.jsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import { HeaderSearch } from '../../components/common/Header/Header';
 import Navbar from '../../components/common/Navbar/Navbar';
 import UserSearch from '../../components/common/User/UserSearch';
-import SearchWrapper from './SearchPageStyle';
 import Layout from "../../styles/Layout";
+import styled from "styled-components";
 function SearchPage() {
   const [data, setdata] = useState(null);
   return (
@@ -16,3 +16,13 @@ function SearchPage() {
 }
 
 export default SearchPage;
+
+const SearchWrapper = styled.section`
+  height: 88vh;
+  width: inherit;
+  /* padding: 16px 24px; */
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  overflow-y: scroll;
+`;
