--- src/pages/SearchPage/SearchPageStyle.jsx
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-
-const SearchWrapper = styled.section`
-  height: 88vh;
-  width: inherit;
-  /* padding: 16px 24px; */
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  overflow-y: scroll;
-`;
-
-export default SearchWrapper;
